requirements.txt
Flask==2.3.2
flask-mysqldb==1.0.1
PyMySQL==1.1.0
python-dotenv==1.0.0
